+++ DroneUpgradeStation/Configs.cs
@@ -12,6 +12,7 @@ namespace DroneUpgradeStation
         public static ConfigEntry<int> minStage { get; private set; }
         public static ConfigEntry<string> sceneList { get; private set; }
         public static ConfigEntry<int> weight { get; private set; }
+        public static ConfigEntry<bool> spawnOnComputationalExchange { get; private set; }
 
         public static string defaultBlacklist = "TreasureCache, FreeChest, Firework, Squid, LowerPricedChests, ExtraShrineItem, Feather, JumpBoost, Duplicator, BarrageOnBoss, ExtraEquipment";
         public static string defaultSceneList = "dampcavesimple, shipgraveyard, rootjungle, repurposedcrater, conduitcanyon, skymeadow, helminthroost";
@@ -27,6 +28,7 @@ namespace DroneUpgradeStation
             minStage = cfg.Bind("Spawn Parameters", "Minimum Stage", 1, "Set the earliest stage number it can spawn in.");
             sceneList = cfg.Bind("Spawn Parameters", "Scene List", defaultSceneList, "List the internal names of the stages it'll spawn in.\nMake sure to separate them with commas.");
             weight = cfg.Bind("Spawn Parameters", "Selection Weight", 4, "Set the odds of it being selected to spawn.\nBy default, it is as rare as Drone Combiner Stations, which is to say, pretty uncommon.");
+            spawnOnComputationalExchange = cfg.Bind("Spawn Parameters", "Spawn on Computational Exchange", true, "A guaranteed Drone Upgrade Station will appear on Computational Exchange.");
         }
     }
 }