+++ DroneUpgradeStation/Main.cs
@@ -7,6 +7,9 @@ using UnityEngine.AddressableAssets;
 using System.Linq;
 using MonoMod.Cil;
 using System;
+using UnityEngine.UI;
+using UnityEngine.Networking;
+using LeTai.Asset.TranslucentImage;
 
 namespace DroneUpgradeStation
 {
@@ -14,7 +17,7 @@ namespace DroneUpgradeStation
     [BepInDependency(DirectorAPI.PluginGUID)]
 
     // Metadata
-    [BepInPlugin("Samuel17.DroneUpgradeStation", "DroneUpgradeStation", "1.0.0")]
+    [BepInPlugin("Samuel17.DroneUpgradeStation", "DroneUpgradeStation", "1.0.1")]
 
     public class Main : BaseUnityPlugin
     {
@@ -23,6 +26,7 @@ namespace DroneUpgradeStation
         public static Material matDroneAssemblyStation = Addressables.LoadAssetAsync<Material>("RoR2/DLC3/DroneAssemblyStation/matDroneAssemblyStation.mat").WaitForCompletion();
         public static GameObject droneAssemblyStationPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC3/DroneAssemblyStation/DroneAssemblyStation.prefab").WaitForCompletion();
         public static ExplicitPickupDropTable blacklist = Addressables.LoadAssetAsync<ExplicitPickupDropTable>("RoR2/DLC3/DroneAssemblyStation/ExcludedItemsDropTable.asset").WaitForCompletion();
+        public static GameObject pickerPanel = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC3/DroneAssemblyStation/AssemblyStationPickerPanelv2.prefab").WaitForCompletion();
         public static ExpansionDef dlc3 = Addressables.LoadAssetAsync<ExpansionDef>("RoR2/DLC3/DLC3.asset").WaitForCompletion();
 
         public void Awake()
@@ -83,7 +87,7 @@ namespace DroneUpgradeStation
                         Log.Error("Drone Upgrade Station BodyFlag hook (2/2) failed!");
                     }
                 };
-            } 
+            }
 
             // Lock to DLC3
             ExpansionRequirementComponent expansionRequirementComponent = droneAssemblyStationPrefab.GetComponent<ExpansionRequirementComponent>();
@@ -95,6 +99,12 @@ namespace DroneUpgradeStation
 
             // Adjust model color
             matDroneAssemblyStation.color = new Color(128f/255f, 255f/255f, 0f/255f);
+
+            // Guaranteed spawn on Computational Exchange
+            if (Configs.spawnOnComputationalExchange.Value == true)
+            {
+                Stage.onStageStartGlobal += OnStageStartGlobal;
+            }
 
             // Functionality
             OtherStuff();
@@ -126,6 +136,21 @@ namespace DroneUpgradeStation
             };
         }
 
+        private void OnStageStartGlobal(Stage stage)
+        {
+            if (SceneInfo.instance.sceneDef != SceneCatalog.FindSceneDef("computationalexchange"))
+            {
+                return;
+            }
+
+            Vector3 position = new Vector3(-47.6f, 121f, -50f);
+            Vector3 rotation = new Vector3(5f, 165f, 0f);
+
+            GameObject station = Instantiate(droneAssemblyStationPrefab, position, Quaternion.Euler(rotation));
+            station.transform.eulerAngles = rotation;
+            NetworkServer.Spawn(station);
+        }
+
         private void OtherStuff()
         {
             // Remove unnecessary PurchaseInteraction; it works fine without it.
@@ -133,6 +158,29 @@ namespace DroneUpgradeStation
             if (purchaseInteract)
             {
                 Utils.RemoveComponent<PurchaseInteraction>(droneAssemblyStationPrefab);
+            }
+
+            // Improve UI
+            On.RoR2.UI.PickupPickerPanel.SetPickupOptions += PickupPickerPanel_SetPickupOptions;
+
+            RectTransform rt = pickerPanel.GetComponent<RectTransform>();
+            if (rt)
+            {
+                rt.localScale = new Vector3(.75f, .75f, .75f);
+                rt.anchorMax = new Vector2(1.1f, 1f);
+            }
+            
+            pickerPanel.GetComponent<TranslucentImage>().enabled = false;
+
+            RoR2.UI.PickupPickerPanel ppp = pickerPanel.GetComponent<RoR2.UI.PickupPickerPanel>();
+            if (ppp)
+            {
+                ppp.maxColumnCount = 12;
+                GridLayoutGroup glg = ppp.gridlayoutGroup;
+                if (glg)
+                {
+                    glg.GetComponent<RectTransform>().localScale = new Vector3(.8f, .8f, .8f);
+                }
             }
         }
 
@@ -189,6 +237,53 @@ namespace DroneUpgradeStation
                         AddToBlacklist(itemDef);
                     }
                 }
+            }
+        }
+
+        private void PickupPickerPanel_SetPickupOptions(On.RoR2.UI.PickupPickerPanel.orig_SetPickupOptions orig, RoR2.UI.PickupPickerPanel self, PickupPickerController.Option[] options) // BROUGHT TO YOU BY LOOKINGGLASS
+        {
+            if (!self.name.StartsWith("AssemblyStationPickerPanelv2"))
+            {
+                orig(self, options);
+                return;
+            }
+
+            Transform t = self.transform.Find("MainPanel");
+            if (t is not null)
+            {
+                Transform background = t.Find("Juice/BG");
+                if (background is not null)
+                {
+                    Color originalColor = background.GetComponent<Image>().color;
+                    background.GetComponent<Image>().color = new Color(originalColor.r, originalColor.g, originalColor.b, .5f);
+                }
+            }
+
+            int itemCount = options.Length;
+
+            int maxHeight = 12;
+            int value = Mathf.CeilToInt((Mathf.Sqrt(itemCount) + 2));
+            GridLayoutGroup gridLayoutGroup = self.transform.GetComponentInChildren<GridLayoutGroup>();
+            if (gridLayoutGroup)
+            {
+                gridLayoutGroup.constraint = GridLayoutGroup.Constraint.FixedColumnCount;
+                gridLayoutGroup.constraintCount = 12;
+                self.maxColumnCount = gridLayoutGroup.constraintCount;
+                self.automaticButtonNavigation = true;
+            }
+
+            orig(self, options);
+
+            if (t is not null)
+            {
+                RectTransform r = t.GetComponent<RectTransform>();
+
+                float height = Mathf.Min(value, maxHeight) * (r.sizeDelta.x / 8f);
+                value = value <= maxHeight ? value : value + 1 + value - maxHeight;
+                float width = (value) * (r.sizeDelta.x / 8f);
+                width = Mathf.Max(width, 340f);
+                height = Mathf.Max(height, 340f);
+                r.sizeDelta = new Vector2(width, height);
             }
         }
     }