+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- A guaranteed Drone Upgrade Station will show up in Computational Exchange. Can you find it?
+- Stole code from LookingGlass to make the UI less ass to navigate.
+
 ## 1.0.0
 
 - Release.